Struct tokio_graceful_shutdown::Toplevel

pub struct Toplevel<ErrType: ErrTypeTraits = Box<dyn Error + Send + Sync + 'static>> { /* private fields */ }

Acts as the base for the subsystem tree and forms the entry point for any interaction with this crate.

Every project that uses this crate has to create a Toplevel object somewhere.

Examples

use miette::Result;
use tokio::time::Duration;
use tokio_graceful_shutdown::{SubsystemHandle, Toplevel};

async fn my_subsystem(subsys: SubsystemHandle) -> Result<()> {
    subsys.request_shutdown();
    Ok(())
}

#[tokio::main]
async fn main() -> Result<()> {
    // Create toplevel
    Toplevel::new()
        .start("MySubsystem", my_subsystem)
        .catch_signals()
        .handle_shutdown_requests(Duration::from_millis(1000))
        .await
        .map_err(Into::into)
}

Implementations

impl<ErrType: ErrTypeTraits> Toplevel<ErrType>

pub fn new() -> Self

Creates a new Toplevel object.

The Toplevel object is the base for everything else in this crate.

pub fn nested(parent: &SubsystemHandle, name: &str) -> Self

Creates a new nested Toplevel object.

This method is identical to .new(), except that the returned Toplevel object will receive shutdown requests from the given SubsystemHandle object.

Any errors caused by subsystems inside the new Toplevel object will cause the Toplevel object to initiate a shutdown, but will not propagate up to the SubsystemHandle object.

Arguments

parent - The subsystemhandle that the Toplevel object will receive shutdown requests from

name - The name of the nested toplevel object. Can be "".

pub fn start<Err, Fut, Subsys>(self, name: &str, subsystem: Subsys) -> Selfwhere
Subsys: 'static + FnOnce(SubsystemHandle<ErrType>) -> Fut + Send,
Fut: 'static + Future<Output = Result<(), Err>> + Send,
Err: Into<ErrType>,

Starts a new subsystem.

Once called, the subsystem will be started immediately, similar to tokio::spawn.

Subsystem

The functionality of the subsystem is represented by the ‘subsystem’ argument. It has to be provided either as an asynchronous function or an asynchronous closure.

It gets provided with a SubsystemHandle object which can be used to interact with this crate.

Returns

When the subsystem returns Ok(()) it is assumed that the subsystem was stopped intentionally and no further actions are performed.

When the subsystem returns an Err, it is assumed that the subsystem failed and a program shutdown gets initiated.

Arguments

name - The name of the subsystem

subsystem - The subsystem to be started

pub fn catch_signals(self) -> Self

Registers signal handlers to initiate a program shutdown when certain operating system signals get received.

The following signals will be handled:

On Windows:

Ctrl+C (SIGINT)

On Unix:

SIGINT and SIGTERM

Caveats

This function internally uses tokio::signal with all of its caveats.

Especially the caveats from tokio::signal::unix::Signal are important for Unix targets.

pub async fn handle_shutdown_requests(
self,
shutdown_timeout: Duration
) -> Result<(), GracefulShutdownError<ErrType>>

Performs a clean program shutdown, once a shutdown is requested or all subsystems have finished.

In most cases, this will be the final method of main(), as it blocks until program shutdown and returns an appropriate Result that can be directly returned by main().

When a program shutdown happens, this function collects the return values of all subsystems to determine the return code of the entire program.

When the shutdown takes longer than the given timeout, an error will be returned and remaining subsystems will be cancelled.

Arguments

shutdown_timeout - The maximum time that is allowed to pass after a shutdown was initiated.

Returns

An error of type GracefulShutdownError if an error occurred.