Trait tokio_graceful_shutdown::IntoSubsystem

pub trait IntoSubsystem<Err, ErrWrapper = Box<dyn Error + Send + Sync + 'static>>where
    Self: Sized + Send + Sync + 'static,
    Err: Into<ErrWrapper>,
    ErrWrapper: ErrTypeTraits,{
    fn run<'async_trait>(
        self,
        subsys: SubsystemHandle<ErrWrapper>
    ) -> Pin<Box<dyn Future<Output = Result<(), Err>> + Send + 'async_trait>>
    where
        Self: 'async_trait;

    fn into_subsystem(
        self
    ) -> Box<dyn FnOnce(SubsystemHandle<ErrWrapper>) -> Pin<Box<dyn Future<Output = Result<(), Err>> + Send + 'static>> + Send + 'static> { ... }
}

Allows a struct to be used as a subsystem.

Implementing this trait requires the async_trait dependency.

Using a struct that does not implement this trait as a subsystem is possible by wrapping it in an async closure. This trait exists primarily for convenience.

The template parameter of the trait is the error type that the subsytem returns.

Examples

use miette::Result;
use tokio::time::Duration;
use tokio_graceful_shutdown::{IntoSubsystem, SubsystemHandle, Toplevel};

struct MySubsystem;

#[async_trait::async_trait]
impl IntoSubsystem<miette::Report> for MySubsystem {
    async fn run(self, subsys: SubsystemHandle) -> Result<()> {
        subsys.request_shutdown();
        Ok(())
    }
}

#[tokio::main]
async fn main() -> Result<()> {
    // Create toplevel
    Toplevel::new()
        .start("Subsys1", MySubsystem{}.into_subsystem())
        .catch_signals()
        .handle_shutdown_requests(Duration::from_millis(500))
        .await
        .map_err(Into::into)
}

Required Methods

fn run<'async_trait>(
self,
subsys: SubsystemHandle<ErrWrapper>
) -> Pin<Box<dyn Future<Output = Result<(), Err>> + Send + 'async_trait>>where
Self: 'async_trait,

The logic of the subsystem.

Will be called as soon as the subsystem gets started.

Returning an error automatically initiates a shutdown.

For more information about subsystem functions, see Toplevel::start() and SubsystemHandle::start().

Provided Methods

fn into_subsystem(
self
) -> Box<dyn FnOnce(SubsystemHandle<ErrWrapper>) -> Pin<Box<dyn Future<Output = Result<(), Err>> + Send + 'static>> + Send + 'static>

Converts the object into a type that can be passed into Toplevel::start() and SubsystemHandle::start().